+++ src/button.h
@@ -16,13 +16,13 @@ class button
 {
 protected:
   IoPi * myIoPi;
-  uint32_t pinUp;
-  uint32_t pinDown;
+  uint8_t pinUp;
+  uint8_t pinDown;
   uint32_t timerStuck;
   char const *name;
   buttonEvents_t lastEvent;
 public:
-  button(char const *name, IoPi *io, uint32_t u, uint32_t d);
+  button(char const *name, IoPi *io, uint8_t u, uint8_t d);
   buttonEvents_t getSignal();
 };
